move network util functions from main.go to their own file in relayer
